% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lava.nlin-package.R
\docType{package}
\name{lava.nlin-package}
\alias{lava.nlin-package}
\alias{lava.nlin}
\title{Non-linear structural equation models.}
\description{
Estimation of parameters in non-linear structural equation models via MCEM,
Laplace Approximation or Importance Sampling
}
\author{
Klaus K. Holst <k.k.holst@biostat.ku.dk>
}
\keyword{package}
