\name{Eval}
\alias{Eval}
\alias{Lapl}
\alias{Mstep_nsem1}
\alias{Mstep_nsem1b}
\alias{Mstep_weibullmm}
\alias{as.mcmc.StEM}
\alias{coef.StEM}
\alias{hessian.weibull}
\alias{finalize}
\alias{lap}
\alias{logl.weibull}
\alias{merge.StEM}
\alias{obj.weibull}
\alias{plot.StEM}
\alias{print.StEM}
\alias{restart}
\alias{restart.StEM}
\alias{score.weibull}
\alias{sim.StEM}
\alias{sim.phwreg}
\alias{weibull.maximize}
\alias{weibullmm}
\alias{window.StEM}


\title{For internal use}
\description{for internal use}
\author{Klaus K. Holst}
\keyword{utilities}
