\name{lava.nlin-package}
\alias{lava.nlin-package}
\alias{lava.nlin}
\docType{package}
\title{
  Non-linear structural equation models. 
}
\description{
  Estimation of parameters in non-linear structural equation models via
  MCEM, Laplace Approximation or Importance Sampling
}
\details{
\tabular{ll}{
Package: \tab lava.nlin\cr
Type: \tab Package\cr
Version: \tab 0.2-4\cr
Date: \tab 2011-10-05\cr
License: \tab GPLv3\cr
LazyLoad: \tab yes\cr
}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
  Klaus K. Holst <k.k.holst@biostat.ku.dk>
}
\keyword{ package }
